requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3
boto3==1.34.162
openai==1.51.2
pydantic==2.9.2
asyncio
google-generativeai